llm/gemini/python-simple/pyproject.toml
[project]
name = "python-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.2.2",
    "ruff>=0.8.3",
]
